Redact the console fallback on read
The transcript was redacted and the console was not, which mattered more
than a coverage gap: the follower stops before kill(), so shutdown output
— panics, the last thing a guest says — landed only in the unredacted
file. The moment most worth reading was the one least protected, on the
path a detached run resolves to.

The obvious instruction, "run the same PiiRedactor over console bytes",
could not be followed. PiiRedactor implements mvm-hostd's egress
Inspector and carries async_trait, InspectorVerdict and RequestCtx, but
the console readers are in mvm-core and are used by mvm-client too —
both below mvm-hostd. Injecting a trait from mvm-cli would have covered
the CLI and left mvm-client reading raw bytes.

So the detector moves down instead. Rules, validators and the RegexSet
machinery have no inspector dependency and now live in mvm_core::pii;
mvm-hostd keeps the Inspector impl wrapping it, which the orphan rule
allows because the trait is local there. mvm-core already carried regex,
so nothing new is linked. The detector's own tests move with it and the
inspector's stay behind. PiiRedactor's fields stayed private, so the
verdict mapping now goes through the mode() accessor.

ConsoleTail holds the redactor non-optionally: a tail that shows raw
bytes is the gap being closed, so it should not be constructible.
Redaction is read-side, so the file the VMM owns is untouched and no
write access is needed. Default rules rather than the workload's
resolved policy, because this path answers a VM whose plan may be long
gone.

One limit, pinned by a test rather than left to be discovered: chunking
is by read, so a value straddling a 64 KiB boundary is two partial
matches and neither fires. Carrying a tail would close it and withhold
the newest bytes from a follower until more arrived — a worse defect on
the path operators reach for when nothing else works. The operator
notices, the module docs, ADR-035 and the public guide all stop saying
the console is unredacted and start naming the boundary instead.
