+++ crates/mvm-cli/src/commands/vm/logs.rs
@@ -405,9 +405,10 @@ fn announce_console_only(args: &Args, sinks: &mut Sinks<'_>) {
     notice(
         sinks,
         &format!(
-            "note: microVM {:?} has no output capture; showing its console log, which is not \
-             redacted, merges stdout and stderr, is not hash-chained, and has no record \
-             boundaries — so -n is approximated in bytes",
+            "note: microVM {:?} has no output capture; showing its console log, which \
+             merges stdout and stderr, is not hash-chained, and has no record boundaries \
+             — so -n is approximated in bytes, and a value split across a 64 KiB read is \
+             not redacted",
             args.name
         ),
     );
@@ -418,17 +419,16 @@ fn announce_console_only(args: &Args, sinks: &mut Sinks<'_>) {
 ///
 /// The recorded half was sealed by a process that did not capture it, so it
 /// stops where *that* process exited — for a detached start, seconds into the
-/// run. Everything after it comes off the console log: same bytes the guest
-/// wrote, none of the guarantees — including redaction, which the recorded
-/// half applied and this half does not. Saying so is what stops the unchained
-/// remainder being read as if the chain covered it, and what explains the
-/// repeated lines where the two halves overlap.
+/// run. Everything after it comes off the console log: redacted like the
+/// recorded half, but with none of the chain's guarantees. Saying so is what
+/// stops the unchained remainder being read as if the chain covered it, and
+/// what explains the repeated lines where the two halves overlap.
 fn describe_history_and_console(args: &Args) -> String {
     format!(
         "note: microVM {:?} was recorded by a process that exited before the run did; what the \
-         recording covers is shown first, then the console log covering the rest — which is not \
-         redacted, merges stdout and stderr, is not hash-chained, and repeats the part the \
-         recording already showed",
+         recording covers is shown first, then the console log covering the rest — which \
+         merges stdout and stderr, is not hash-chained, and repeats the part the recording \
+         already showed",
         args.name
     )
 }
@@ -782,7 +782,10 @@ mod tests {
         // The spliced-in half is raw guest bytes. Losing the chain costs
         // accuracy; losing redaction is the one with a safety consequence, so
         // it is named rather than left to the guide.
-        assert!(described.contains("not redacted"), "{described}");
+        assert!(
+            !described.contains("which is not redacted"),
+            "the console half is redacted now; the notice must not still deny it: {described}"
+        );
     }
 
     #[test]
@@ -1346,10 +1349,12 @@ mod tests {
             let rendered = notices(captured);
             assert!(rendered.contains("merges stdout and stderr"), "{rendered}");
             assert!(rendered.contains("not hash-chained"), "{rendered}");
-            // The console log is the unredacted source. An operator reading it
-            // as if the broker's redaction applied is the one misreading here
-            // that has a safety consequence, so the notice names it.
-            assert!(rendered.contains("not redacted"), "{rendered}");
+            // The console is redacted now, so the notice must not claim
+            // otherwise — an operator who reads "not redacted" and goes
+            // hunting for a rawer source is being sent somewhere that does
+            // not exist. What stays named is what is still true: merged
+            // channels, no chain, and the read-boundary limit.
+            assert!(!rendered.contains("which is not redacted"), "{rendered}");
             assert!(
                 rendered.contains("-n is approximated in bytes"),
                 "{rendered}"
